+++ src/Products/MeetingCommunes/tests/testCustomMeeting.py
@@ -137,3 +137,90 @@ class testCustomMeeting(MeetingCommunesTestCase):
             meeting.adapted().getNumberOfItems(
                 itemUids, categories=['development', ], listTypes=['normal']),
             0)
+
+    def test_GetPrintableItemsByCategoryWithBothLateItems(self):
+        self.changeUser('pmManager')
+        self.setMeetingConfig(self.meetingConfig2.getId())
+        meeting = self._createMeetingWithItems()
+        orderedItems = meeting.getItems(ordered=True)
+        item1 = orderedItems[0]
+        item2 = orderedItems[1]
+        item3 = orderedItems[2]
+        self.do(item1, 'backToValidated')
+        self.do(item1, 'backToProposed')
+        self.do(item2, 'backToValidated')
+        self.do(item2, 'backToProposed')
+        self.do(item3, 'backToValidated')
+        self.do(item3, 'backToProposed')
+        self.freezeMeeting(meeting)
+        item1.setPreferredMeeting(meeting.UID())
+        item2.setPreferredMeeting(meeting.UID())
+        item3.setPreferredMeeting(meeting.UID())
+        self.presentItem(item1)
+        self.presentItem(item2)
+        self.presentItem(item3)
+        # now we have 2 normal items and 3 late items
+        # 2 lates development, 1 normal and 1 late events
+        # and 1 normal research
+        # build the list of uids
+        itemUids = [anItem.UID() for anItem in meeting.getItems(ordered=True)]
+        # test on the meeting with listTypes=['late','normal']
+        # Every items (normal and late) should be in the same category, in the good order
+        self.assertEqual(
+            meeting.adapted().getPrintableItemsByCategory(
+                itemUids, listTypes=['late', 'normal'])[0][0].getId(),
+            'development')
+        self.assertEqual(
+            meeting.adapted().getPrintableItemsByCategory(
+                itemUids, listTypes=['late', 'normal'])[1][0].getId(),
+            'events')
+        self.assertEqual(
+            meeting.adapted().getPrintableItemsByCategory(
+                itemUids, listTypes=['late', 'normal'])[2][0].getId(),
+            'research')
+        self.assertEqual(
+            meeting.adapted().getPrintableItemsByCategory(
+                itemUids, listTypes=['late', 'normal'])[0][0].portal_type,
+            'meetingcategory')
+        self.assertEqual(
+            meeting.adapted().getPrintableItemsByCategory(
+                itemUids, listTypes=['late', 'normal'])[1][0].portal_type,
+            'meetingcategory')
+        self.assertEqual(
+            meeting.adapted().getPrintableItemsByCategory(
+                itemUids, listTypes=['late', 'normal'])[2][0].portal_type,
+            'meetingcategory')
+        # the event category should have 2 items, research 1 and development 2 ( + 1 category element for each one)
+        self.assertEqual(
+            len(meeting.adapted().getPrintableItemsByCategory(
+                itemUids, listTypes=['late', 'normal'])[0]),
+            3)
+        self.assertEqual(
+            len(meeting.adapted().getPrintableItemsByCategory(
+                itemUids, listTypes=['late', 'normal'])[1]),
+            3)
+        self.assertEqual(
+            len(meeting.adapted().getPrintableItemsByCategory(
+                itemUids, listTypes=['late', 'normal'])[2]),
+            2)
+        # other element of the list are MeetingItems...
+        self.assertEqual(
+            meeting.adapted().getPrintableItemsByCategory(
+                itemUids, listTypes=['late', 'normal'])[0][1].meta_type,
+            'MeetingItem')
+        self.assertEqual(
+            meeting.adapted().getPrintableItemsByCategory(
+                itemUids, listTypes=['late', 'normal'])[0][2].meta_type,
+            'MeetingItem')
+        self.assertEqual(
+            meeting.adapted().getPrintableItemsByCategory(
+                itemUids, listTypes=['late', 'normal'])[1][1].meta_type,
+            'MeetingItem')
+        self.assertEqual(
+            meeting.adapted().getPrintableItemsByCategory(
+                itemUids, listTypes=['late', 'normal'])[1][2].meta_type,
+            'MeetingItem')
+        self.assertEqual(
+            meeting.adapted().getPrintableItemsByCategory(
+                itemUids, listTypes=['late', 'normal'])[2][1].meta_type,
+            'MeetingItem')